+++ src/packages/string_basic.rs
@@ -399,7 +399,6 @@ mod number_formatting {
         }
 
         #[cfg(not(target_family = "wasm"))]
-
         pub mod num_128 {
             /// Convert the `value` into a string in hex format.
             #[rhai_fn(name = "to_hex")]